enc,dec
This script implements the AES (Advanced Encryption Standard) encryption algorithm in Python. It includes the basic AES encryption functions along with the ability to work in different modes of operation, such as ECB, CBC, CFB, OFB, and CTR. The script also demonstrates the padding and key expansion methods necessary for AES encryption.